+++ src/common/MovieBanner/Ratings/styled.js
