+++ src/common/MovieBanner/styled.js
