+++ src/common/SectionBanner/styled.js
